requirements.txt
beautifulsoup4==4.12.3
Requests==2.32.3
typing_extensions==4.12.2
google-generativeai==0.7.2